Version 1.0.1
Added support for marking extraneous whitespace not followed by an element. Since (sigh) ABNF seems to be whitespace sensitive, that could be a problem.

Added support for marking comments which are not part of a rule (i.e. at the beginning of a file).

Both of these new features may be enabled or disabled according to your preference.

The whitespace matching is enabled by default, but may be disabled via {{{let abnf_ignore_whitespace=1}}}.

The comments thing is disabled by default and may be enabled via {{{let abnf_strict=1}}}